requirements.txt
alembic==1.4.3
attrs==21.2.0
Babel==2.9.1
blinker==1.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Cors==3.0.9
Flask-JWT-Extended==4.2.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.15.1
idna==3.2
importlib-metadata==4.6.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.5.0
